Aomp_vrc_tmp/requirements/requirements.txt
absl-py==0.6.1
aiohttp==3.4.4
aiohttp-cors==0.7.0
aniso8601==4.0.1
asn1crypto==0.24.0
astor==0.7.1
async-timeout==3.0.1
attrs==18.2.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
common==0.1.2
cryptography==2.4.2
cycler==0.10.0
dask==0.20.1
decorator==4.3.0
easydict==1.9
Flask==1.0.2
Flask-Cors==3.0.7
Flask-Extension==1.0
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
gast==0.2.0
gevent==1.3.7
greenlet==0.4.15
grpcio==1.16.0
gunicorn==19.9.0
h5py==2.8.0
idna==2.7
idna-ssl==1.1.0
itsdangerous==1.1.0
Jinja2==2.10
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
kiwisolver==1.0.1
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.1
msgpack==0.6.0
msgpack-numpy==0.4.4.2
multidict==4.5.2
networkx==2.2
numpy==1.14.5
opencv-contrib-python==3.2.0.7
opencv-python==3.4.4.19
Pillow==5.3.0
protobuf==3.6.1
pybase64==0.4.0
pycparser==2.19
PyMySQL==0.9.2
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
PyWavelets==1.0.1
PyYAML==3.13
pyzmq==17.1.2
requests==2.20.1
scikit-image==0.14.1
scikit-learn==0.20.2
scipy==1.1.0
six==1.11.0
sklearn==0.0
slim==0.3.14
SQLAlchemy==1.2.14
tabulate==0.8.2
tensorboard==1.10.0
tensorflow==1.10.0
tensorpack==0.9.0.1
termcolor==1.1.0
toolz==0.9.0
tqdm==4.28.1
urllib3==1.24.1
virtualenv==16.1.0
Werkzeug==0.14.1
wincertstore==0.2
xmltodict==0.11.0
yarl==1.3.0
